Re #2407 Move GlobalOptsMonoid out of Stack.Types.Config
