+++ src/Options/Applicative/Complicated.hs
@@ -29,7 +29,8 @@ import           Options.Applicative.Builder.Internal
                    ( Mod (..), mkCommand, mkParser )
 import           Options.Applicative.Types ( OptReader (..) )
 import           Stack.Prelude
-import           Stack.Types.Config ( AddCommand, GlobalOptsMonoid )
+import           Stack.Types.AddCommand ( AddCommand )
+import           Stack.Types.GlobalOptsMonoid ( GlobalOptsMonoid )
 import           Stack.Types.Runner ( Runner )
 import           System.Environment ( getArgs )
