+++ src/Stack/Config.hs
@@ -103,19 +103,17 @@ import           Stack.Types.Build ( BuildException (..), FlagSource (..) )
 import           Stack.Types.ColorWhen ( ColorWhen (..) )
 import           Stack.Types.Compiler ( defaultCompilerRepository )
 import           Stack.Types.Config
-                   ( BuildConfig (..)
-                   , BuildOpts (..), Config (..)
-                   , ConfigException (..)
-                   , ConfigPrettyException (..), HasConfig (..)
-                   , ParseAbsolutePathException (..)
+                   ( BuildConfig (..), BuildOpts (..), Config (..)
+                   , ConfigException (..), ConfigPrettyException (..)
+                   , HasConfig (..), ParseAbsolutePathException (..)
                    , PlatformVariant (..), Project (..)
                    , ProjectAndConfigMonoid (..), ProjectConfig (..)
                    , askLatestSnapshotUrl, configProjectRoot
-                   , packageIndicesWarning, parseConfigMonoid
-                   , parseProjectAndConfigMonoid, platformOnlyRelDir
-                   , stackRootL, workDirL
+                   , packageIndicesWarning, parseProjectAndConfigMonoid
+                   , platformOnlyRelDir, stackRootL, workDirL
                    )
-import           Stack.Types.ConfigMonoid ( ConfigMonoid (..) )
+import           Stack.Types.ConfigMonoid
+                   ( ConfigMonoid (..), parseConfigMonoid )
 import           Stack.Types.Docker ( DockerOptsMonoid (..), dockerEnable )
 import           Stack.Types.DumpLogs ( DumpLogs (..) )
 import           Stack.Types.GlobalOpts (  GlobalOpts (..) )
