+++ src/Stack/Types/Config.hs
@@ -51,17 +51,12 @@ module Stack.Types.Config
   , ConfigPrettyException (..)
   , ParseAbsolutePathException (..)
   , packageIndicesWarning
-  -- ** ConfigMonoid
-  , configMonoidInstallGHCName
-  , configMonoidSystemGHCName
-  , parseConfigMonoid
   -- ** EnvSettings
   , EnvSettings (..)
   , minimalEnvSettings
   , defaultEnvSettings
   , plainEnvSettings
   -- ** GlobalOpts & GlobalOptsMonoid
-  , GlobalOptsMonoid (..)
   , defaultLogLevel
   -- ** Project & ProjectAndConfigMonoid
   , Project (..)
@@ -95,7 +90,6 @@ module Stack.Types.Config
   , workDirL
   , ghcInstallHook
   -- * Command-related types
-  , AddCommand
   , module X
   -- * Lens helpers
   , wantedCompilerVersionL
@@ -130,26 +124,17 @@ module Stack.Types.Config
   , to
   ) where
 
-import           Control.Monad.Writer ( Writer, tell )
-import           Control.Monad.Trans.Except ( ExceptT )
 import           Crypto.Hash ( SHA1 (..), hashWith )
 import           Pantry.Internal.AesonExtended
-                   ( FromJSON, FromJSONKey (..)
-                   , FromJSONKeyFunction (FromJSONKeyTextParser), Object, ToJSON
-                   , Value, WarningParser, WithJSONWarnings (..), (.=)
-                   ,  (...:), (..:?), (..!=), jsonSubWarnings
-                   , jsonSubWarningsT, jsonSubWarningsTT, object, parseJSON
-                   , toJSON, withObjectWarnings, withText
+                   ( FromJSON (..), ToJSON (..), Value, WithJSONWarnings (..)
+                   , (.=), (...:), (..:?), (..!=), jsonSubWarnings
+                   , jsonSubWarningsT, jsonSubWarningsTT, object
+                   , withObjectWarnings
                    )
-import           Data.Attoparsec.Args ( EscapingMode (Escaping), parseArgs )
 import qualified Data.ByteArray.Encoding as Mem ( Base(Base16), convertToBase )
 import qualified Data.ByteString.Char8 as S8
-import           Data.Coerce ( coerce )
 import qualified Data.List.NonEmpty as NonEmpty
 import qualified Data.Map as Map
-import qualified Data.Map.Strict as M
-import qualified Data.Monoid as Monoid
-import           Data.Monoid.Map ( MonoidMap (..) )
 import qualified Data.Set as Set
 import qualified Data.Text as T
 import           Data.Yaml ( ParseException )
@@ -160,7 +145,7 @@ import           Distribution.PackageDescription ( GenericPackageDescription )
 import qualified Distribution.PackageDescription as C
 import           Distribution.System ( Platform, Arch )
 import qualified Distribution.Text ( display )
-import           Distribution.Version ( anyVersion, mkVersion )
+import           Distribution.Version ( mkVersion )
 import           Generics.Deriving.Monoid ( mappenddefault, memptydefault )
 import           Lens.Micro ( _1, _2 )
 import           Options.Applicative ( ReadM )
@@ -171,6 +156,7 @@ import           Path
                    ( (</>), dirname, filename, parent, parseAbsDir, parseAbsFile
                    , parseRelDir, parseRelFile, reldir, relfile
                    )
+import           RIO.Process ( HasProcessContext (..), ProcessContext )
 import           Stack.Constants
                    ( bindirSuffix, docDirSuffix, ghcColorForceFlag, osIsWindows
                    , relDirCompilerTools, relDirHoogle, relDirHpc, relDirInstall
@@ -185,15 +171,15 @@ import           Stack.Types.Compiler
                    )
 import           Stack.Types.CompilerBuild
                    ( CompilerBuild, compilerBuildSuffix )
-import           Stack.Types.ConfigMonoid ( ConfigMonoid (..) )
-import           Stack.Types.Docker
-                   ( DockerOpts, VersionRangeJSON (..) )
-import           Stack.Types.DockerEntrypoint ( DockerEntrypoint )
+import           Stack.Types.ConfigMonoid
+                   ( ConfigMonoid (..), configMonoidAllowDifferentUserName
+                   , configMonoidGHCVariantName, configMonoidSystemGHCName
+                   , parseConfigMonoidObject)
+import           Stack.Types.Docker ( DockerOpts )
 import           Stack.Types.DumpLogs ( DumpLogs )
 import           Stack.Types.GHCVariant ( GHCVariant (..), ghcVariantSuffix )
 import           Stack.Types.GhcPkgId ( GhcPkgId )
 import           Stack.Types.GlobalOpts ( GlobalOpts (..) )
-import           Stack.Types.LockFileBehavior ( LockFileBehavior )
 import           Stack.Types.NamedComponent ( NamedComponent (..) )
 import           Stack.Types.Nix ( NixOpts )
 import           Stack.Types.PvpBounds ( PvpBounds )
@@ -207,11 +193,9 @@ import           Stack.Types.SourceMap
                    )
 import           Stack.Types.TemplateName ( TemplateName )
 import           Stack.Types.Version
-                   ( IntersectingVersionRange (..), VersionCheck (..)
-                   , VersionRange, stackVersion, versionRangeText
+                   ( VersionCheck (..), VersionRange, stackVersion
+                   , versionRangeText
                    )
-import qualified System.FilePath as FilePath
-import           RIO.Process ( HasProcessContext (..), ProcessContext )
 
 -- Re-exports
 import           Stack.Types.Config.Build as X
@@ -640,12 +624,6 @@ data EnvSettings = EnvSettings
   }
   deriving (Eq, Ord, Show)
 
-type AddCommand =
-  ExceptT (RIO Runner ())
-          (Writer (OA.Mod OA.CommandFields (RIO Runner (), GlobalOptsMonoid)))
-          ()
-
-
 -- | Project configuration information. Not every run of Stack has a
 -- true local project; see constructors below.
 data ProjectConfig a
@@ -657,47 +635,6 @@ data ProjectConfig a
     -- the implicit global.
   | PCNoProject ![PackageIdentifierRevision]
     -- ^ Use a no project run. This comes from 'SYLNoProject'.
-
--- | Parsed global command-line options monoid.
-data GlobalOptsMonoid = GlobalOptsMonoid
-  { globalMonoidReExecVersion :: !(First String)
-    -- ^ Expected re-exec in container version
-  , globalMonoidDockerEntrypoint :: !(First DockerEntrypoint)
-    -- ^ Data used when Stack is acting as a Docker entrypoint (internal use
-    -- only)
-  , globalMonoidLogLevel     :: !(First LogLevel)
-    -- ^ Log level
-  , globalMonoidTimeInLog    :: !FirstTrue
-    -- ^ Whether to include timings in logs.
-  , globalMonoidRSLInLog     :: !FirstFalse
-    -- ^ Whether to include raw snapshot layer (RSL) in logs.
-  , globalMonoidConfigMonoid :: !ConfigMonoid
-    -- ^ Config monoid, for passing into 'loadConfig'
-  , globalMonoidResolver     :: !(First (Unresolved AbstractResolver))
-    -- ^ Resolver override
-  , globalMonoidResolverRoot :: !(First FilePath)
-    -- ^ root directory for resolver relative path
-  , globalMonoidCompiler     :: !(First WantedCompiler)
-    -- ^ Compiler override
-  , globalMonoidTerminal     :: !(First Bool)
-    -- ^ We're in a terminal?
-  , globalMonoidStyles       :: !StylesUpdate
-    -- ^ Stack's output styles
-  , globalMonoidTermWidth    :: !(First Int)
-    -- ^ Terminal width override
-  , globalMonoidStackYaml    :: !(First FilePath)
-    -- ^ Override project stack.yaml
-  , globalMonoidLockFileBehavior :: !(First LockFileBehavior)
-    -- ^ See 'globalLockFileBehavior'
-  }
-  deriving Generic
-
-instance Semigroup GlobalOptsMonoid where
-  (<>) = mappenddefault
-
-instance Monoid GlobalOptsMonoid where
-  mempty = memptydefault
-  mappend = (<>)
 
 -- | Default logging level should be something useful but not crazy.
 defaultLogLevel :: LogLevel
@@ -847,335 +784,6 @@ instance FromJSON (WithJSONWarnings Curator) where
     <*> fmap (Set.map unCabalString) (o ..:? "expect-benchmark-failure" ..!= mempty)
     <*> fmap (Set.map unCabalString) (o ..:? "skip-haddock" ..!= mempty)
     <*> fmap (Set.map unCabalString) (o ..:? "expect-haddock-failure" ..!= mempty)
-
-parseConfigMonoid ::
-     Path Abs Dir
-  -> Value
-  -> Yaml.Parser (WithJSONWarnings ConfigMonoid)
-parseConfigMonoid = withObjectWarnings "ConfigMonoid" . parseConfigMonoidObject
-
--- | Parse a partial configuration.  Used both to parse both a standalone config
--- file and a project file, so that a sub-parser is not required, which would
--- interfere with warnings for missing fields.
-parseConfigMonoidObject :: Path Abs Dir -> Object -> WarningParser ConfigMonoid
-parseConfigMonoidObject rootDir obj = do
-  -- Parsing 'stackRoot' from 'stackRoot'/config.yaml would be nonsensical
-  let configMonoidStackRoot = First Nothing
-  configMonoidWorkDir <- First <$> obj ..:? configMonoidWorkDirName
-  configMonoidBuildOpts <-
-    jsonSubWarnings (obj ..:? configMonoidBuildOptsName ..!= mempty)
-  configMonoidDockerOpts <-
-    jsonSubWarnings (obj ..:? configMonoidDockerOptsName ..!= mempty)
-  configMonoidNixOpts <-
-    jsonSubWarnings (obj ..:? configMonoidNixOptsName ..!= mempty)
-  configMonoidConnectionCount <-
-    First <$> obj ..:? configMonoidConnectionCountName
-  configMonoidHideTHLoading <-
-    FirstTrue <$> obj ..:? configMonoidHideTHLoadingName
-  configMonoidPrefixTimestamps <-
-    First <$> obj ..:? configMonoidPrefixTimestampsName
-
-  murls :: Maybe Value <- obj ..:? configMonoidUrlsName
-  configMonoidLatestSnapshot <-
-    case murls of
-      Nothing -> pure $ First Nothing
-      Just urls -> jsonSubWarnings $ lift $ withObjectWarnings
-        "urls"
-        (\o -> First <$> o ..:? "latest-snapshot" :: WarningParser (First Text))
-        urls
-
-  configMonoidPackageIndex <-
-    First <$> jsonSubWarningsT (obj ..:?  configMonoidPackageIndexName)
-  configMonoidPackageIndices <-
-    First <$> jsonSubWarningsTT (obj ..:?  configMonoidPackageIndicesName)
-  configMonoidSystemGHC <- First <$> obj ..:? configMonoidSystemGHCName
-  configMonoidInstallGHC <- FirstTrue <$> obj ..:? configMonoidInstallGHCName
-  configMonoidSkipGHCCheck <-
-    FirstFalse <$> obj ..:? configMonoidSkipGHCCheckName
-  configMonoidSkipMsys <- FirstFalse <$> obj ..:? configMonoidSkipMsysName
-  configMonoidRequireStackVersion <-
-    IntersectingVersionRange . unVersionRangeJSON <$>
-      ( obj ..:? configMonoidRequireStackVersionName
-          ..!= VersionRangeJSON anyVersion
-      )
-  configMonoidArch <- First <$> obj ..:? configMonoidArchName
-  configMonoidGHCVariant <- First <$> obj ..:? configMonoidGHCVariantName
-  configMonoidGHCBuild <- First <$> obj ..:? configMonoidGHCBuildName
-  configMonoidJobs <- First <$> obj ..:? configMonoidJobsName
-  configMonoidExtraIncludeDirs <- map (toFilePath rootDir FilePath.</>) <$>
-    obj ..:?  configMonoidExtraIncludeDirsName ..!= []
-  configMonoidExtraLibDirs <- map (toFilePath rootDir FilePath.</>) <$>
-    obj ..:?  configMonoidExtraLibDirsName ..!= []
-  configMonoidCustomPreprocessorExts <-
-    obj ..:?  configMonoidCustomPreprocessorExtsName ..!= []
-  configMonoidOverrideGccPath <-
-    First <$> obj ..:? configMonoidOverrideGccPathName
-  configMonoidOverrideHpack <-
-    First <$> obj ..:? configMonoidOverrideHpackName
-  configMonoidConcurrentTests <-
-    First <$> obj ..:? configMonoidConcurrentTestsName
-  configMonoidLocalBinPath <- First <$> obj ..:? configMonoidLocalBinPathName
-  templates <- obj ..:? "templates"
-  (configMonoidScmInit,configMonoidTemplateParameters) <-
-    case templates of
-      Nothing -> pure (First Nothing,M.empty)
-      Just tobj -> do
-        scmInit <- tobj ..:? configMonoidScmInitName
-        params <- tobj ..:? configMonoidTemplateParametersName
-        pure (First scmInit,fromMaybe M.empty params)
-  configMonoidCompilerCheck <-
-    First <$> obj ..:? configMonoidCompilerCheckName
-  configMonoidCompilerRepository <-
-    First <$> (obj ..:? configMonoidCompilerRepositoryName)
-
-  options <-
-    Map.map unGhcOptions <$> obj ..:? configMonoidGhcOptionsName ..!= mempty
-
-  optionsEverything <-
-    case (Map.lookup GOKOldEverything options, Map.lookup GOKEverything options) of
-      (Just _, Just _) ->
-        fail "Cannot specify both `*` and `$everything` GHC options"
-      (Nothing, Just x) -> pure x
-      (Just x, Nothing) -> do
-        tell "The `*` ghc-options key is not recommended. Consider using \
-             \$locals, or if really needed, $everything"
-        pure x
-      (Nothing, Nothing) -> pure []
-
-  let configMonoidGhcOptionsByCat = coerce $ Map.fromList
-        [ (AGOEverything, optionsEverything)
-        , (AGOLocals, Map.findWithDefault [] GOKLocals options)
-        , (AGOTargets, Map.findWithDefault [] GOKTargets options)
-        ]
-
-      configMonoidGhcOptionsByName = coerce $ Map.fromList
-          [(name, opts) | (GOKPackage name, opts) <- Map.toList options]
-
-  configMonoidCabalConfigOpts' <- obj ..:? "configure-options" ..!= mempty
-  let configMonoidCabalConfigOpts =
-        coerce (configMonoidCabalConfigOpts' :: Map CabalConfigKey [Text])
-
-  configMonoidExtraPath <- obj ..:? configMonoidExtraPathName ..!= []
-  configMonoidSetupInfoLocations <-
-    obj ..:? configMonoidSetupInfoLocationsName ..!= []
-  configMonoidSetupInfoInline <-
-    jsonSubWarningsT (obj ..:? configMonoidSetupInfoInlineName) ..!= mempty
-  configMonoidLocalProgramsBase <-
-    First <$> obj ..:? configMonoidLocalProgramsBaseName
-  configMonoidPvpBounds <- First <$> obj ..:? configMonoidPvpBoundsName
-  configMonoidModifyCodePage <-
-    FirstTrue <$> obj ..:? configMonoidModifyCodePageName
-  configMonoidRebuildGhcOptions <-
-    FirstFalse <$> obj ..:? configMonoidRebuildGhcOptionsName
-  configMonoidApplyGhcOptions <-
-    First <$> obj ..:? configMonoidApplyGhcOptionsName
-  configMonoidAllowNewer <- First <$> obj ..:? configMonoidAllowNewerName
-  configMonoidAllowNewerDeps <- obj ..:? configMonoidAllowNewerDepsName
-  configMonoidDefaultTemplate <-
-    First <$> obj ..:? configMonoidDefaultTemplateName
-  configMonoidAllowDifferentUser <-
-    First <$> obj ..:? configMonoidAllowDifferentUserName
-  configMonoidDumpLogs <- First <$> obj ..:? configMonoidDumpLogsName
-  configMonoidSaveHackageCreds <-
-    First <$> obj ..:? configMonoidSaveHackageCredsName
-  configMonoidHackageBaseUrl <-
-    First <$> obj ..:? configMonoidHackageBaseUrlName
-
-  configMonoidColorWhenUS <- obj ..:? configMonoidColorWhenUSName
-  configMonoidColorWhenGB <- obj ..:? configMonoidColorWhenGBName
-  let configMonoidColorWhen =  First $   configMonoidColorWhenUS
-                                     <|> configMonoidColorWhenGB
-
-  configMonoidStylesUS <- obj ..:? configMonoidStylesUSName
-  configMonoidStylesGB <- obj ..:? configMonoidStylesGBName
-  let configMonoidStyles = fromMaybe mempty $   configMonoidStylesUS
-                                            <|> configMonoidStylesGB
-
-  configMonoidHideSourcePaths <-
-    FirstTrue <$> obj ..:? configMonoidHideSourcePathsName
-  configMonoidRecommendUpgrade <-
-    FirstTrue <$> obj ..:? configMonoidRecommendUpgradeName
-
-  configMonoidCasaRepoPrefix <-
-    First <$> obj ..:? configMonoidCasaRepoPrefixName
-  configMonoidSnapshotLocation <-
-    First <$> obj ..:? configMonoidSnapshotLocationName
-  configMonoidNoRunCompile <-
-    FirstFalse <$> obj ..:? configMonoidNoRunCompileName
-
-  configMonoidStackDeveloperMode <-
-    First <$> obj ..:? configMonoidStackDeveloperModeName
-
-  pure ConfigMonoid {..}
-
-configMonoidWorkDirName :: Text
-configMonoidWorkDirName = "work-dir"
-
-configMonoidBuildOptsName :: Text
-configMonoidBuildOptsName = "build"
-
-configMonoidDockerOptsName :: Text
-configMonoidDockerOptsName = "docker"
-
-configMonoidNixOptsName :: Text
-configMonoidNixOptsName = "nix"
-
-configMonoidConnectionCountName :: Text
-configMonoidConnectionCountName = "connection-count"
-
-configMonoidHideTHLoadingName :: Text
-configMonoidHideTHLoadingName = "hide-th-loading"
-
-configMonoidPrefixTimestampsName :: Text
-configMonoidPrefixTimestampsName = "build-output-timestamps"
-
-configMonoidUrlsName :: Text
-configMonoidUrlsName = "urls"
-
-configMonoidPackageIndexName :: Text
-configMonoidPackageIndexName = "package-index"
-
--- Deprecated in favour of package-index
-configMonoidPackageIndicesName :: Text
-configMonoidPackageIndicesName = "package-indices"
-
-configMonoidSystemGHCName :: Text
-configMonoidSystemGHCName = "system-ghc"
-
-configMonoidInstallGHCName :: Text
-configMonoidInstallGHCName = "install-ghc"
-
-configMonoidSkipGHCCheckName :: Text
-configMonoidSkipGHCCheckName = "skip-ghc-check"
-
-configMonoidSkipMsysName :: Text
-configMonoidSkipMsysName = "skip-msys"
-
-configMonoidRequireStackVersionName :: Text
-configMonoidRequireStackVersionName = "require-stack-version"
-
-configMonoidArchName :: Text
-configMonoidArchName = "arch"
-
-configMonoidGHCVariantName :: Text
-configMonoidGHCVariantName = "ghc-variant"
-
-configMonoidGHCBuildName :: Text
-configMonoidGHCBuildName = "ghc-build"
-
-configMonoidJobsName :: Text
-configMonoidJobsName = "jobs"
-
-configMonoidExtraIncludeDirsName :: Text
-configMonoidExtraIncludeDirsName = "extra-include-dirs"
-
-configMonoidExtraLibDirsName :: Text
-configMonoidExtraLibDirsName = "extra-lib-dirs"
-
-configMonoidCustomPreprocessorExtsName  :: Text
-configMonoidCustomPreprocessorExtsName  = "custom-preprocessor-extensions"
-
-configMonoidOverrideGccPathName :: Text
-configMonoidOverrideGccPathName = "with-gcc"
-
-configMonoidOverrideHpackName :: Text
-configMonoidOverrideHpackName = "with-hpack"
-
-configMonoidConcurrentTestsName :: Text
-configMonoidConcurrentTestsName = "concurrent-tests"
-
-configMonoidLocalBinPathName :: Text
-configMonoidLocalBinPathName = "local-bin-path"
-
-configMonoidScmInitName :: Text
-configMonoidScmInitName = "scm-init"
-
-configMonoidTemplateParametersName :: Text
-configMonoidTemplateParametersName = "params"
-
-configMonoidCompilerCheckName :: Text
-configMonoidCompilerCheckName = "compiler-check"
-
-configMonoidCompilerRepositoryName :: Text
-configMonoidCompilerRepositoryName = "compiler-repository"
-
-configMonoidGhcOptionsName :: Text
-configMonoidGhcOptionsName = "ghc-options"
-
-configMonoidExtraPathName :: Text
-configMonoidExtraPathName = "extra-path"
-
-configMonoidSetupInfoLocationsName :: Text
-configMonoidSetupInfoLocationsName = "setup-info-locations"
-
-configMonoidSetupInfoInlineName :: Text
-configMonoidSetupInfoInlineName = "setup-info"
-
-configMonoidLocalProgramsBaseName :: Text
-configMonoidLocalProgramsBaseName = "local-programs-path"
-
-configMonoidPvpBoundsName :: Text
-configMonoidPvpBoundsName = "pvp-bounds"
-
-configMonoidModifyCodePageName :: Text
-configMonoidModifyCodePageName = "modify-code-page"
-
-configMonoidRebuildGhcOptionsName :: Text
-configMonoidRebuildGhcOptionsName = "rebuild-ghc-options"
-
-configMonoidApplyGhcOptionsName :: Text
-configMonoidApplyGhcOptionsName = "apply-ghc-options"
-
-configMonoidAllowNewerName :: Text
-configMonoidAllowNewerName = "allow-newer"
-
-configMonoidAllowNewerDepsName :: Text
-configMonoidAllowNewerDepsName = "allow-newer-deps"
-
-configMonoidDefaultTemplateName :: Text
-configMonoidDefaultTemplateName = "default-template"
-
-configMonoidAllowDifferentUserName :: Text
-configMonoidAllowDifferentUserName = "allow-different-user"
-
-configMonoidDumpLogsName :: Text
-configMonoidDumpLogsName = "dump-logs"
-
-configMonoidSaveHackageCredsName :: Text
-configMonoidSaveHackageCredsName = "save-hackage-creds"
-
-configMonoidHackageBaseUrlName :: Text
-configMonoidHackageBaseUrlName = "hackage-base-url"
-
-configMonoidColorWhenUSName :: Text
-configMonoidColorWhenUSName = "color"
-
-configMonoidColorWhenGBName :: Text
-configMonoidColorWhenGBName = "colour"
-
-configMonoidStylesUSName :: Text
-configMonoidStylesUSName = "stack-colors"
-
-configMonoidStylesGBName :: Text
-configMonoidStylesGBName = "stack-colours"
-
-configMonoidHideSourcePathsName :: Text
-configMonoidHideSourcePathsName = "hide-source-paths"
-
-configMonoidRecommendUpgradeName :: Text
-configMonoidRecommendUpgradeName = "recommend-stack-upgrade"
-
-configMonoidCasaRepoPrefixName :: Text
-configMonoidCasaRepoPrefixName = "casa-repo-prefix"
-
-configMonoidSnapshotLocationName :: Text
-configMonoidSnapshotLocationName = "snapshot-location-base"
-
-configMonoidNoRunCompileName :: Text
-configMonoidNoRunCompileName = "script-no-run-compile"
-
-configMonoidStackDeveloperModeName :: Text
-configMonoidStackDeveloperModeName = "stack-developer-mode"
 
 packageIndicesWarning :: StyleDoc
 packageIndicesWarning =
@@ -1513,36 +1121,6 @@ data PlatformVariant
 platformVariantSuffix :: PlatformVariant -> String
 platformVariantSuffix PlatformVariantNone = ""
 platformVariantSuffix (PlatformVariant v) = "-" ++ v
-
-data GhcOptionKey
-  = GOKOldEverything
-  | GOKEverything
-  | GOKLocals
-  | GOKTargets
-  | GOKPackage !PackageName
-  deriving (Eq, Ord)
-
-instance FromJSONKey GhcOptionKey where
-  fromJSONKey = FromJSONKeyTextParser $ \t ->
-    case t of
-      "*" -> pure GOKOldEverything
-      "$everything" -> pure GOKEverything
-      "$locals" -> pure GOKLocals
-      "$targets" -> pure GOKTargets
-      _ ->
-        case parsePackageName $ T.unpack t of
-          Nothing -> fail $ "Invalid package name: " ++ show t
-          Just x -> pure $ GOKPackage x
-  fromJSONKeyList =
-    FromJSONKeyTextParser $ \_ -> fail "GhcOptionKey.fromJSONKeyList"
-
-newtype GhcOptions = GhcOptions { unGhcOptions :: [Text] }
-
-instance FromJSON GhcOptions where
-  parseJSON = withText "GhcOptions" $ \t ->
-    case parseArgs Escaping t of
-      Left e -> fail e
-      Right opts -> pure $ GhcOptions $ map T.pack opts
 
 -----------------------------------
 -- Lens classes
