+++ src/main/Stack/CLI.hs
@@ -67,8 +67,9 @@ import           Stack.SDist ( sdistCmd )
 import           Stack.Script ( ScriptOpts (..), scriptCmd )
 import           Stack.SetupCmd ( setupCmd )
 import           Stack.Templates ( templatesCmd )
-import           Stack.Types.Config
-                   ( AddCommand, BuildCommand (..), GlobalOptsMonoid (..) )
+import           Stack.Types.AddCommand ( AddCommand )
+import           Stack.Types.Config ( BuildCommand (..) )
+import           Stack.Types.GlobalOptsMonoid ( GlobalOptsMonoid (..) )
 import           Stack.Types.Runner ( Runner )
 import           Stack.Types.Version ( stackVersion )
 import           Stack.Uninstall ( uninstallCmd )
